+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './MasterFisher';
 export * from './Mixin';
 export * from './InstanceDefinitionUtils';
 export * from './Processing';
+export * from './PathUtils';
